# storage/api-docs/api-docs.yaml
openapi: 3.0.0
info:
  title: 'The Financial Express API'
  contact:
    name: 'Eliyas Hossain'
    url: 'https://batterylowinteractive.com'
    email: eliyas.batterylowinteractive@gmail.com
  version: 1.0.0
servers:
  -
    url: /admin
    description: 'Admin API base url'
paths:
...
  /auth/login:
    post:
      tags:
        - Auth
      description: 'Login by email, password'
      operationId: authLogin
      requestBody:
        description: 'Pass user credentials'
        required: true
        content:
          application/json:
            schema:
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                  example: user@example.com
                password:
                  type: string
                  format: password
                  example: PassWord12345
                remember:
                  type: boolean
                  example: 'true'
              type: object
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                    example: Success
                type: object
        400:
          description: 'Bad Request'
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                    example: 'Something weng wrong, please try again later.'
                type: object
        422:
          description: 'Wrong credentials response'
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                    example: 'Something weng wrong, please try again later.'
                type: object
  /auth/token:
    post:
      tags:
        - Auth
      description: 'Generate Token to use API docs'
      operationId: authToken
      requestBody:
        description: 'Pass user credentials'
        required: true
        content:
          application/json:
            schema:
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                  example: user@example.com
                password:
                  type: string
                  format: password
                  example: PassWord12345
              type: object
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                    example: Success
                type: object
        400:
          description: 'Bad Request'
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                    example: 'Something weng wrong, please try again later.'
                type: object
        422:
          description: 'Wrong credentials response'
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                    example: 'Something weng wrong, please try again later.'
                type: object
  /auth/me:
    get:
      tags:
        - Auth
      description: 'Get loggedIn user'
      operationId: authMe
      responses:
        200:
          description: 'Authentication response'
          content:
            application/json:
              schema:
                properties:
                  user:
                    $ref: '#/components/schemas/AuthResponse'
                type: object
        400:
          description: 'Bad Request'
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                    example: 'Something weng wrong, please try again later.'
                type: object
        401:
          description: Unauthenticated
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                    example: 'Please sign in to use this service.'
                type: object
      security:
        -
          sanctum: []
  /auth/logout:
    post:
      tags:
        - Auth
      description: 'Logout and desctory session'
      operationId: authLogout
      responses:
        200:
          description: 'Successfully Logged Out'
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                    example: 'Successfully Logged Out!'
                type: object
        400:
          description: 'Bad Request'
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                    example: 'Something weng wrong, please try again later.'
                type: object
        401:
          description: Unauthenticated
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                    example: 'Please sign in to use this service.'
                type: object
      security:
        -
          sanctum: []
  /auth/logout-by-token:
    post:
      tags:
        - Auth
      description: 'Logout and delete token'
      operationId: authLogoutByToken
      responses:
        200:
          description: 'Successfully Logged Out'
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                    example: 'Successfully Logged Out!'
                type: object
        400:
          description: 'Bad Request'
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                    example: 'Something weng wrong, please try again later.'
                type: object
        401:
          description: Unauthenticated
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                    example: 'Please sign in to use this service.'
                type: object
      security:
        -
          sanctum: []
  /auth/forgot-password:
    post:
      tags:
        - Auth
      description: 'Forgot password request'
      operationId: authForgotPassword
      requestBody:
        description: 'Please enter email address'
        required: true
        content:
          application/json:
            schema:
              required:
                - email
              properties:
                email:
                  type: string
                  format: email
                  example: user@example.com
              type: object
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                    example: Success
                type: object
        400:
          description: 'Bad Request'
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                    example: 'Something weng wrong, please try again later.'
                type: object
        422:
          description: 'Validation Error Response'
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                    example: 'Something weng wrong, please try again later.'
                type: object
  /auth/reset-password:
    post:
      tags:
        - Auth
      description: 'Reset password request'
      operationId: authResetPassword
      requestBody:
        description: 'Please enter valid information'
        required: true
        content:
          application/json:
            schema:
              required:
                - token
                - email
                - password
                - password_confimation
              properties:
                token:
                  type: string
                  format: string
                  example: 1231231232321kjkljkljklk
                email:
                  type: string
                  format: email
                  example: user@example.com
                password:
                  type: string
                  format: string
                  example: PassWord12345
                password_confirmation:
                  type: string
                  format: string
                  example: PassWord12345
              type: object
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                    example: Success
                type: object
        400:
          description: 'Bad Request'
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                    example: 'Something weng wrong, please try again later.'
                type: object
  '/auth/email/verify/{id}/{hash}':
    post:
      tags:
        - Auth
      description: 'Verify email address'
      operationId: authVerifyEmail
      parameters:
        -
          name: id
          in: path
          description: 'ID of User'
          required: true
          schema:
            type: integer
            format: int64
          example: '1'
        -
          name: hash
          in: path
          description: 'Hash value'
          required: true
          schema:
            type: string
          example: 12345asdfg
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                    example: Success
                type: object
        404:
          description: 'Not Found'
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                    example: 'Not Found.'
                type: object
        400:
          description: 'Bad Request'
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                    example: 'Something weng wrong, please try again later.'
                type: object
  '/auth/email/resend/{email}':
    post:
      tags:
        - Auth
      description: 'Resend verification email'
      operationId: authResendVerifyEmail
      parameters:
        -
          name: email
          in: path
          description: 'Email address'
          required: true
          schema:
            type: string
            format: email
          example: user@example.com
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                    example: Success
                type: object
        404:
          description: 'Not Found'
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                    example: 'Not Found.'
                type: object
        400:
          description: 'Bad Request'
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                    example: 'Something weng wrong, please try again later.'
                type: object
...
components:
  schemas:
...
    AuthResponse:
      title: AuthResponse
      description: 'Authentication response'
      properties:
        id:
          title: 'User ID'
          description: ID
          type: integer
          format: int64
        name:
          title: Name
          description: 'User name'
          type: string
          format: string
        email:
          title: Email
          description: 'User email'
          type: string
          format: email
        mobile:
          title: Mobile
          description: 'User mobile'
          type: string
          format: string
        avatar:
          title: Avatar
          description: 'User avatar'
          type: string
          format: string
        role:
          title: Role
          description: 'User role'
          type: string
          format: string
      type: object
...
  securitySchemes:
    sanctum:
      type: http
      description: 'Laravel Sanctum token authentication'
      scheme: bearer
      bearerFormat: JWT
security:
  - []